requirements.txt
streamlit==1.29.0
pandas==2.0.3
requests==2.31.0
cryptography==41.0.3
websockets==11.0.3
python-dotenv==1.0.0
kalshi-python==3.2.1
dateparser==1.1.8
jupyter==1.0.0
ipykernel==6.25.1
pytest==7.4.0
